requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0